hypersentry/backend/requirements.txt
# Core Framework
fastapi
starlette
python-multipart

# Async HTTP
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
bitarray
certifi
cffi
charset-normalizer
ckzg
colorlog
cryptography
cytoolz
eth-account
eth-hash
eth-keyfile
eth-keys
eth-rlp
eth-typing
eth-utils
eth_abi
frozenlist
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-genai
googleapis-common-protos
grpcio
grpcio-status
h11
hexbytes
httpcore
httplib2
httpx
hyperliquid-python-sdk
idna
msgpack
multidict
parsimonious
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pycryptodome
pydantic
pydantic_core
pyparsing
python-dotenv
python-telegram-bot
regex
python-dateutil
requests
rlp
toolz
tqdm
typing-inspection
typing_extensions
uritemplate
urllib3
websocket-client
yarl
gunicorn
uvicorn

celery
redis

# Database ORM
sqlalchemy
alembic
psycopg2-binary

# JWT Authentication
python-jose[cryptography]

# HTTP client for OAuth
httpx
pandas
numpy
websockets
